add Style and search bar
